+++ client/src/ciaoUtils/ciaoDbg.ts
@@ -1,9 +1,8 @@
 'use strict';
 
 import type { DebugMark } from '../../../shared/types';
+import { dbgMarkRegex } from '../constants';
 import { translatePath } from './ciaoFile';
-
-const DBG_MARK_REGEX = / {9}In (.*) \(([0-9]+)-([0-9]+)\) (.*?)-([0-9]+)/g;
 
 /**
  * Determines wether the line is a debugging line or not
@@ -11,8 +10,18 @@ const DBG_MARK_REGEX = / {9}In (.*) \(([0-9]+)-([0-9]+)\) (.*?)-([0-9]+)/g;
  * @returns `true` if it is, `false` otherwise
  */
 export function isDebuggerLine(line: string): boolean {
-  return DBG_MARK_REGEX.test(line);
+  return dbgMarkRegex.test(line);
 }
+
+/*
+ * Useful regexes?
+ *
+ *     [/^   [0-9]+  [0-9]+  Call: /m, 'comment'],
+ *     [/^   [0-9]+  [0-9]+  Exit: /m, 'comment'],
+ *     [/^   [0-9]+  [0-9]+  Redo: /m, 'comment'],
+ *     [/^   [0-9]+  [0-9]+  Fail: /m, 'comment'],
+ *
+ */
 
 /**
  * Parses a debugger message and returns the information
@@ -20,7 +29,7 @@ export function isDebuggerLine(line: string): boolean {
  * @returns Object containing all the extracted information
  */
 export function parseDbgMsg(msg: string): DebugMark | undefined {
-  const [match] = [...msg.matchAll(DBG_MARK_REGEX)];
+  const match = msg.match(dbgMarkRegex);
   if (!match) return;
   const [, srcFile, startLine, endLine, predName, nthPred] = match;
   return {