+++ client/src/utils.ts
@@ -4,7 +4,7 @@ import * as os from 'node:os';
 import { mkdtempSync, readdirSync } from 'node:fs';
 import path, { join } from 'node:path';
 import { spawn, spawnSync } from 'node:child_process';
-import { commands, window, workspace } from 'vscode';
+import { Uri, Webview, commands, window, workspace } from 'vscode';
 import { getGlobalValue, setGlobalValue } from './contextManager';
 import { ciaoInstallerCmd } from './constants';
 import { type OS, type CiaoVersion } from '../../shared/types';
@@ -242,6 +242,39 @@ export async function reloadVSCode(): Promise<void> {
   }
 }
 
+/**
+ * A helper function that returns a unique alphanumeric identifier called a nonce.
+ *
+ * @remarks This function is primarily used to help enforce content security
+ * policies for resources/scripts being executed in a webview context.
+ *
+ * @returns A nonce
+ */
+export function getNonce(): string {
+  let text = '';
+  const possible =
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  for (let i = 0; i < 32; i++) {
+    text += possible.charAt(Math.floor(Math.random() * possible.length));
+  }
+  return text;
+}
+
+/**
+ * A helper function which will get the webview URI of a given file or resource.
+ *
+ * @remarks This URI can be used within a webview's HTML as a link to the
+ * given file/resource.
+ *
+ * @param webview A reference to the extension webview
+ * @param extensionUri The URI of the directory containing the extension
+ * @param pathList An array of strings representing the path to a file/resource
+ * @returns A URI pointing to the file/resource
+ */
+export function getUri(webview: Webview, extensionUri: Uri, pathList: string[]) {
+  return webview.asWebviewUri(Uri.joinPath(extensionUri, ...pathList));
+}
+
 /*
  * @returns Path of the Ciao Version inside the PATH
  */
@@ -253,8 +286,8 @@ function getCiaoVersionFromPath(): CiaoVersion | undefined {
 
   return stdout
     ? {
-        name: 'PATH',
-        path: stdout.toString().split(path.sep).slice(0, -3).join(path.sep),
-      }
+      name: 'PATH',
+      path: stdout.toString().split(path.sep).slice(0, -3).join(path.sep),
+    }
     : undefined;
 }